managed_services/requirements.txt
# Managed services requirements
# Core dependencies
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
pydantic>=2.5.0

# Cloud service SDKs
boto3>=1.34.0                    # AWS Textract
google-cloud-documentai>=2.20.0  # Google Document AI
azure-ai-formrecognizer>=3.3.0   # Azure Form Recognizer
azure-core>=1.29.0

# AI/ML dependencies
google-generativeai>=0.3.0      # Gemini API

# Authentication
PyJWT>=2.8.0                    # JWT token handling

# Google Cloud (for prompt caching)
google-auth>=2.23.0             # Google Cloud authentication
google-auth-oauthlib>=1.1.0     # OAuth2 flow
google-auth-httplib2>=0.2.0     # HTTP library for auth
requests>=2.31.0                # HTTP requests

# Document processing (lightweight versions)
PyMuPDF>=1.23.0                 # PDF processing
python-docx>=1.1.0              # Word documents
Pillow>=10.1.0                  # Image processing

# Existing app dependencies (for compatibility)
# Note: These should already be installed in your environment
# fastapi
# uvicorn
# python-multipart
# azure-data-tables
# azure-storage-blob

# Development and testing
pytest>=7.4.0
httpx>=0.25.0                   # For testing FastAPI
pytest-asyncio>=0.21.0

# Logging and monitoring
structlog>=23.2.0